social-media/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
django==5.0.4
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0 #OAuth2 
social-auth-core==4.5.1
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
aiogram==3.2.0
python-dotenv==1.0.0
redis==5.0.1
cryptography==41.0.7
Authlib==1.3.0
jwcrypto==1.5.1
django-cors-headers==4.3.1
drf-spectacular==0.27.2
geopy==2.4.1
Pillow
django-cors-headers
python-dotenv